backend/requirements.txt
quart
quart-cors
langchain
fastapi
uvicorn
openai
python-dotenv
pocketbase
PyPDF3
retry